Move tests related to courses' cotes to CoursExtTest
