+++ android/app/src/test/kotlin/ca/etsmtl/applets/etsmobile/presentation/GradesViewModelTest.kt
@@ -5,10 +5,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import ca.etsmtl.applets.etsmobile.R
 import ca.etsmtl.applets.etsmobile.domain.FetchGradesCoursesUseCase
+import ca.etsmtl.applets.etsmobile.extension.mockNetwork
 import ca.etsmtl.applets.etsmobile.presentation.grades.GradesViewModel
 import ca.etsmtl.applets.etsmobile.util.Event
-import ca.etsmtl.applets.etsmobile.extension.adjustCote
-import ca.etsmtl.applets.etsmobile.util.mockNetwork
 import ca.etsmtl.applets.repository.data.model.Resource
 import com.nhaarman.mockito_kotlin.capture
 import com.nhaarman.mockito_kotlin.mock
@@ -148,82 +147,5 @@ class GradesViewModelTest {
 
         // then
         verify(observer).onChanged(false)
-    }
-
-    @Test
-    fun cote_NotNullOrEmpty_ReturnsCote() {
-        // given
-        val cours = Cours(
-            "MAT123",
-            "01",
-            "s.o",
-            "7365",
-            "K",
-            "91",
-            3,
-            "Math"
-        )
-
-        // when
-        val resultCours = with(gradesViewModel) {
-            cours.apply {
-                adjustCote(app)
-            }
-        }
-
-        // then
-        assertEquals(cours, resultCours)
-    }
-
-    @Test
-    fun cote_CoteEmptyAndNotSur100NotNullOrEmpty_ReturnsNoteSur100() {
-        // given
-        val cours = Cours(
-            "MAT123",
-            "01",
-            "s.o",
-            "7365",
-            "",
-            "91",
-            3,
-            "Math"
-        )
-        `when`(app.getString(R.string.text_grade_in_percentage)).thenReturn("%1\$s %%")
-
-        // when
-        val resultCours = with(gradesViewModel) {
-            cours.apply {
-                adjustCote(app)
-            }
-        }
-
-        // then
-        assertEquals(cours.copy(cote = cours.noteSur100 + " %"), resultCours)
-    }
-
-    @Test
-    fun cote_CoteEmptyAndNoteSur100NullOrEmpty_ReturnsNotAvailable() {
-        // given
-        val cours = Cours(
-            "MAT123",
-            "01",
-            "s.o",
-            "7365",
-            "",
-            "",
-            3,
-            "Math"
-        )
-        `when`(app.getString(R.string.abbreviation_not_available)).thenReturn("N/A")
-
-        // when
-        val resultCours = with(gradesViewModel) {
-            cours.apply {
-                adjustCote(app)
-            }
-        }
-
-        // then
-        assertEquals(cours.copy(cote = "N/A"), resultCours)
     }
 }
