+++ android/app/src/test/kotlin/ca/etsmtl/applets/etsmobile/presentation/ProfileViewModelTest.kt
@@ -13,7 +13,7 @@ import ca.etsmtl.applets.etsmobile.presentation.profile.ProfileValueItem
 import ca.etsmtl.applets.etsmobile.presentation.profile.ProfileViewModel
 import ca.etsmtl.applets.etsmobile.util.Event
 import ca.etsmtl.applets.etsmobile.util.EventObserver
-import ca.etsmtl.applets.etsmobile.util.mockNetwork
+import ca.etsmtl.applets.etsmobile.extension.mockNetwork
 import ca.etsmtl.applets.repository.data.model.Resource
 import com.nhaarman.mockito_kotlin.any
 import com.nhaarman.mockito_kotlin.capture
